refactor: Lottos를 List<Lotto>를 주입 받아 생성하도록 변경 - 기존 Strategy를 이용하던 부분을 Machine의 책임으로 이동

+++ src/main/java/enumeration/Rank.java
@@ -55,7 +55,7 @@ public enum Rank {
         return rank;
     }
 
-    public int prize() {
+    public long prize() {
         return prize;
     }
 }
